requirements.txt
# ============================================================
# LULC Classification Pipeline — Dependencies
# For Apple Silicon (M1/M2/M3/M4) Macs
# ============================================================

# TensorFlow for Apple Silicon
tensorflow-macos>=2.14.0
tensorflow-metal>=1.1.0

# Geospatial
rasterio>=1.3.0
earthengine-api>=0.1.370
geemap>=0.30.0

# Core scientific
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0

# Utilities
tqdm>=4.65.0
